Breadth-First Saving
Its been a while, and a few big changes were made, but by far the most impactful functional change was the that to make the saving done using breadth-first instead of inorder (also called level order traversal). This greatly increases the efficiency when loading from the save files as there will be no rotations required when constructing the tree.